Starting the process with breaking down to smaller components and server components
